move unit test in test folder
