+++ app/src/main/java/ch/epfl/sweng/hostme/FragmentCreationPage5.java
@@ -16,6 +16,8 @@ import com.google.firebase.auth.FirebaseAuth;
 import java.util.HashMap;
 import java.util.Map;
 
+import ch.epfl.sweng.hostme.utils.PasswordValidator;
+
 
 public class FragmentCreationPage5 extends Fragment {
     public final static Map<String, String> DATA = new HashMap<>();
