+++ app/src/main/java/ch/epfl/sweng/hostme/utils/EmailValidator.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.hostme;
+package ch.epfl.sweng.hostme.utils;
 
 import android.util.Log;
