+++ app/src/main/java/ch/epfl/sweng/hostme/utils/PasswordValidator.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.hostme;
+package ch.epfl.sweng.hostme.utils;
 
 import java.util.regex.Pattern;
