+++ app/src/test/java/ch/epfl/sweng/hostme/PasswordValidatorTest.java
@@ -4,6 +4,8 @@ import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
+import ch.epfl.sweng.hostme.utils.PasswordValidator;
+
 public class PasswordValidatorTest {
 
     PasswordValidator validator = new PasswordValidator();
